Made Attack functions in tank,mosnter,gunnery + made decHealth takes double not int + added dm parameter in pick functions for oddNum Picked drones + added some directories

+++ Army units/ArmyUnit.cpp
@@ -36,7 +36,7 @@ unitType ArmyUnit::getType()
 	return type;
 }
 
-bool ArmyUnit::DecHlth(int value)
+bool ArmyUnit::DecHlth(double value)
 {
 	hlth -= value;
 	if(hlth<=0)
